requirements.txt
# Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# LLM & AI
openai==1.6.1
anthropic==0.18.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-anthropic==0.0.4
langchain-community==0.0.10

# Document processing
pypdf==3.17.4
python-docx==1.1.0
markdown==3.5.2
beautifulsoup4==4.12.2

# Frontend
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.4

# AWS
boto3==1.34.0
botocore==1.34.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4